data-loader/requirements.txt
pdfplumber==0.11.4
psycopg2-binary==2.9.9
python-dotenv==1.0.1
